requirements-all.txt
# Merged Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
websockets>=12.0
asyncio-mqtt==0.16.1
aioredis==2.0.1
python-socketio>=5.13.0
requests>=2.32.4
urllib3>=2.0.0
websocket-client
eventlet>=0.33.3
aiofiles>=23.1.0
aiohttp>=3.13.3
mss>=10.0.0
dxcam>=0.0.5
opencv-python>=4.7.0.72
numpy>=1.24.4
pillow>=9.5.0
PyTurboJPEG>=1.7.0
pyaudio>=0.2.11
sounddevice>=0.5.2
SpeechRecognition>=3.14.0
pynput>=1.8.1
keyboard>=0.13.5
pyautogui>=0.9.54
pygame>=2.6.1
psutil>=5.9.5
py-cpuinfo>=9.0.0
pywin32>=306
aiortc>=1.13.0
msgpack>=1.0.5
lz4>=4.3.2
zstandard>=0.21.0
xxhash>=3.2.0
cryptography>=41.0.3
Cython>=3.0.0
setuptools>=65.0.0
flask>=2.3.2
flask_socketio>=5.5.1
flask-cors>=4.0.0
gunicorn>=23.0.0
pyotp>=2.8.0
qrcode>=7.4.2
